e2e, decorators: Introduce retain-VMIs decorator
In Kubevirt e2e tests there are tests batches that create the exact same VMI before
each test, where each test assert different thing.
In such cases, reusing the VM in multiple tests can cut down their overall time
and make our CI more efficient.
Also, in enable breaking down long and bloated tests to smaller ones, such tests
cover multiple sceneries which are not necessarily related to each other and should
be extracted to its own test.

The 'RetainVirtualMachineInstances' decorator make the global tests cleanup skip
VMI objects deletion, and by that enable reusing them in other tests.
Eventually, the VMIs will be deleted by the global cleanup on the global test suite
teardown, or earlier, by the next test suite that isn't decorated with retain-VMIs.

How to use:
  Describe("my awesome tests", decorators.RetainVirtualMachineInstances, ...

+++ tests/testsuite/namespace.go
@@ -43,6 +43,7 @@ import (
 	"kubevirt.io/kubevirt/pkg/controller"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
 
+	"kubevirt.io/kubevirt/tests/decorators"
 	"kubevirt.io/kubevirt/tests/flags"
 	"kubevirt.io/kubevirt/tests/framework/checks"
 	"kubevirt.io/kubevirt/tests/framework/cleanup"
@@ -80,6 +81,8 @@ func CleanNamespaces() {
 	virtCli, err := kubecli.GetKubevirtClientFromRESTConfig(restConfig)
 	util.PanicOnError(err)
 
+	currentGinkgoSpecReport := CurrentSpecReport()
+
 	for _, namespace := range TestNamespaces {
 		listOptions := metav1.ListOptions{
 			LabelSelector: cleanup.TestLabelForNamespace(namespace),
@@ -110,9 +113,9 @@ func CleanNamespaces() {
 		// Remove all VirtualMachineReplicaSets
 		util.PanicOnError(virtCli.RestClient().Delete().Namespace(namespace).Resource("virtualmachineinstancereplicasets").Do(context.Background()).Error())
 
-		deleteVirtualMachineInstances(namespace)
+		deleteVirtualMachineInstances(currentGinkgoSpecReport, namespace)
 
-		deletePods(namespace)
+		deletePods(currentGinkgoSpecReport, namespace)
 
 		// Remove all Services
 		svcList, err := virtCli.CoreV1().Services(namespace).List(context.Background(), metav1.ListOptions{})
@@ -228,7 +231,11 @@ func CleanNamespaces() {
 	}
 }
 
-func deleteVirtualMachineInstances(namespace string) {
+func deleteVirtualMachineInstances(specReport SpecReport, namespace string) {
+	if decorators.HasLabel(specReport, decorators.RetainVirtualMachineInstances) {
+		return
+	}
+
 	virtCli := kubevirt.Client()
 	util.PanicOnError(virtCli.RestClient().Delete().Namespace(namespace).Resource("virtualmachineinstances").Do(context.Background()).Error())
 	vmis, err := virtCli.VirtualMachineInstance(namespace).List(context.Background(), &metav1.ListOptions{})
@@ -243,10 +250,18 @@ func deleteVirtualMachineInstances(namespace string) {
 	}
 }
 
-func deletePods(namespace string) {
+func deletePods(specReport SpecReport, namespace string) {
 	virtCli := kubevirt.Client()
-	podList, err := virtCli.CoreV1().Pods(namespace).List(context.Background(), metav1.ListOptions{})
-	util.PanicOnError(err)
+	var podList *k8sv1.PodList
+	var err error
+	if decorators.HasLabel(specReport, decorators.RetainVirtualMachineInstances) {
+		filterOutVirtLauncherPods := fmt.Sprintf("kubevirt.io!=virt-launcher")
+		podList, err = virtCli.CoreV1().Pods(namespace).List(context.Background(), metav1.ListOptions{LabelSelector: filterOutVirtLauncherPods})
+		util.PanicOnError(err)
+	} else {
+		podList, err = virtCli.CoreV1().Pods(namespace).List(context.Background(), metav1.ListOptions{})
+		util.PanicOnError(err)
+	}
 
 	var gracePeriod int64 = 0
 	for _, pod := range podList.Items {
